fix: bug resyncing nonce on error (#36)
Whenever an error is returned after submitting a tx onchain, the inner
NonceManager nonce is reset if it was different from the one onchain.

However, the inner nonce is expected to be greater by 1 than the onchain
nonce in this case, since the new tx is assumed to have called
`self.next()`, which returns the next nonce to use but also increments
the inner nonce by 1.

So previously, resyncing was done when `onchain_nonce != inner_nonce`
This fix does the resyncing when `onchain_none != (inner_nonce - 1)`,
because of the incrementing that happens in `self.next()`

As a drive-by change, this PR also sets the `txs_since_resync` to 0 when
the nonce is resynced due to an error, to avoid over-resyncing.