chore: upgrade `elliptic`, `k256` and `yubi` crates to latest (#40)
## Motivation

https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5625 implements
signing with a YubiHSM2 device in Hyperlane.

Yubihsm signing requires using the `yubihsm` feature from the `yubi`
crate. Both `yubi` and various other parts of `ethers-signers`,
`ethers-core` and `hyperlane-*` libraries require signing code to
implement `DigestSigner`. This interface is closely coupled (AFAICT)
with versions of `elliptic` and `k256`. In otherwords, both seem to need
to be bumped at once.

A [previous reviewer
comment](https://github.com/Tessellated-io/hyperlane-monorepo/pull/2#discussion_r1981769340)
correctly pointed out that for implementing yubihsm2 signing in
hyperlane, we should use a global library version of `k256` and
`elliptic` across the codebase.

As things stand today, we can use the existing versions in `hyperlane`
and pull in the `yubi` crate at v0.41.x. Unfortunately, [there is a
bug](https://github.com/iqlusioninc/yubihsm.rs/pull/464) in 0.41.x <
`yubi` > 0.42.1 which causes a signing request to hang indefinitely
(effectively deadlocking the process) in the case of a network failure.

Bumping `yubi` to v0.42.1 pulls in newer versions of `elliptic` and
`k256` which have replaced the `DigestSigner` interface with a similar
but different `PreHashSigner` interface. I cannot figure out a way to
get `hyperlane` to build with `yubi`@0.42.1 without having a consistent
version of `k256` and `elliptic` across `ethers-rs` and `hyperlane-*`.
Therefore I think that we need to bump these packages as a prerequisite
to https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5625.

If you agree, I can make a corresponding PR to `hyperlane-monorepo` that
shows we can bump these deps there without any code changes.

I ran `clippy` and there are lots of warnings, but I don't think they
are related to changes in the PR (I see them on the main `hyperlane`
branch as well). Likewise, all tests pass on this branch that are
passing on the `hyperlane` branch.


## Solution

This PR is a copy of the upstream PR which upgraded the same deps. See
(upstream pr)[https://github.com/gakonst/ethers-rs/pull/2260].


## PR Checklist

-   [n/a] Added Tests
-   [n/a] Added Documentation
-   [n/a] Updated the changelog